Changes from reviews 1
